% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/startup_functions.R
\name{start_matter}
\alias{start_matter}
\title{Start matter}
\usage{
start_matter(args = commandArgs())
}
\arguments{
\item{args}{optional arguments supplied to establish settings}
}
\description{
A wrapper for a set of functions that should be run at the start
}
\details{
This wrapper controls the following:
\itemize{
\item set the initial stage (Stage 1)
\item initialises the status list
\item loads the necessary packages
\item retrieve any settings from CLA or other
\item defines the paths used across the analyses
\item clear any transient data from various paths
\item ensure all necessary paths are present
}
}
\author{
Murray Logan
}
